Add optional params to form field decoders
Add `min_length` and `max_length` params to list field decoder. This
efficiently counts while traversing and typechecking the list, and
lazily stops as soon as the max list length is surpassed.

Add `default` param to required field decoder. This allows setting a
default value for a required form field.