Simplify path and route types
We only need the type variable used for parsing the route. For printing
the route to an attribute we just need to coerce the type of the format
string to the one needed for the attribute printer. Because the format
strings internally have exactly the same structure, this is sound.

This also allows us to get rid of the `[%path "..."]` PPX, because now
we can construct paths with only a single format string literal.